Update lang.json
This is the Turkish translation from the country Turkey.  I am a native Turkish speaker and I have been living in the US for the past 5 years. I love Tab Space a lot and use it everyday so I just wanted to help you guys grow and keep updating the tab space.